move tests to separate spec
